requirements.txt
Flask==3.0.0
PyYAML==6.0.1
requests==2.31.0
boto3==1.34.0
python-dateutil==2.8.2
pytest==7.4.3
pytest-flask==1.3.0
